{"schema":"decodex/commit/1","summary":"Split Rust review query and MCP planning modules by responsibility","authority":"manual"}
